refactor: 패키지 구조를 domain 과 view 로 분리
- 이에 따라 여러 소스 코드의 접근 제어자 변경
- CarInputTest 는 Cars 테스트에 가까워 클래스 이름을 CarsTest 로 변경
- CycleInputTest 는 Cycle 테스트에 가까워 클래스 이름을 CycleTest 로 변경
