+++ src/main/java/racinggame/domain/CarName.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 public class CarName {
 	private String carName;
