+++ src/main/java/racinggame/domain/Cycle.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 public class Cycle {
 	private int cycle;
