+++ src/main/java/racinggame/view/CycleInput.java
@@ -1,12 +1,13 @@
-package racinggame;
+package racinggame.view;
 
 import java.util.NoSuchElementException;
 
 import nextstep.utils.Console;
+import racinggame.domain.Cycle;
 
 public class CycleInput {
-	static final String GUIDE_MESSAGE = "시도할 회수는 몇회인가요?";
-	static final String ERROR_MESSAGE = "1 이상의 숫자를 입력해주세요.";
+	public static final String GUIDE_MESSAGE = "시도할 회수는 몇회인가요?";
+	public static final String ERROR_MESSAGE = "1 이상의 숫자를 입력해주세요.";
 
 	private CycleInput() {
 	}
