+++ src/test/java/racinggame/domain/CarsTest.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -7,8 +7,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import nextstep.test.NSTest;
+import racinggame.view.CarInput;
 
-public class CarInputTest extends NSTest {
+public class CarsTest extends NSTest {
 	@BeforeEach
 	void beforeEach() {
 		setUp();
